add a note about enabling permalinks
REST API routes don't work with a default installation of Wordpress, which has 'ugly permalinks'. I think it should be mentioned on the first page of the documentation. A new user shouldn't have to hunt on Google for clues on why the api they want to try out doesn't work 'out of the box'. Related issue here: https://github.com/WP-API/WP-API/issues/2623

I am not sure about the wording, so please let me know if you'd phrase it differently.